add abstract activity
